packages/ai-agents/requirements.txt
# Core AI dependencies
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.1.1
openai==1.10.0
anthropic==0.18.0

# Blockchain
web3==6.15.0

# Database & Cache
supabase==2.3.4
httpx==0.27.0

# Utilities
python-dotenv==1.0.0
pydantic==1.10.13
aiohttp==3.9.1
requests==2.31.0
